Clean up and minimize API changes.
This commit moves most of the changes previously made to ScriptRuntime into OptRuntime.
The price we pay for this is some overlap in functionality between OptRuntime and ScriptRuntime,
but we get clean separation between optimized and interpreted code with ScriptRuntime using
only the Scriptable interface to access names and OptRuntime using OptCall directly.

OptCall now implements the Scriptable interface methods which will make it work fine with
interpreted code that doesn't make use of the OptCall API.

This commit also changes behaviour of name lookup and binding in regard to with-statements
with E4X/XML objects. Previously when a name could not be resolved in the scope chain and
there was a with(xmlobj), the name would bind to the XML object. I disabled this because:

 a) I couldn't find any requirement for this behaviour in ECMA 357 and 262.
 b) None of the Mozilla E4X tests require it
 c) Spidermonkey doesn't implement it either